Migrate Statistics screen to Template E and modularize UI
- Refactor StatisticsScreen to use AnalyticsViewTemplate (Template E)
- Reorder content: Chart -> Metrics -> Insights -> Behavior
- Extract UI sections into modular widgets for maintainability
- Enhance ScreenShell with titleWidget and toolbarHeight support
- Localize hardcoded strings and fix linting issues
