+++ lib/features/statistics/presentation/widgets/statistics_insights_section.dart
@@ -1,0 +1,375 @@
+import 'package:flutter/material.dart';
+import 'package:flutter_riverpod/flutter_riverpod.dart';
+import 'package:lucide_icons_flutter/lucide_icons.dart';
+import 'package:mudra_manager/core/currency/currency_meta.dart';
+import 'package:mudra_manager/core/currency/currency_service.dart';
+import 'package:mudra_manager/core/l10n/app_localizations.dart';
+import 'package:mudra_manager/core/providers/spacing_provider.dart';
+import 'package:mudra_manager/features/analytics/data/analytics_provider.dart';
+import 'package:mudra_manager/features/dashboard/data/status_data_provider.dart';
+import 'package:mudra_manager/shared/widgets/currency_text.dart';
+import 'package:mudra_manager/shared/widgets/inline_error.dart';
+import 'package:visibility_detector/visibility_detector.dart';
+
+class StatisticsInsightsSection extends ConsumerWidget {
+  final StatsData data;
+
+  const StatisticsInsightsSection({super.key, required this.data});
+
+  @override
+  Widget build(BuildContext context, WidgetRef ref) {
+    final spacing = ref.watch(spacingProvider);
+    final color = Theme.of(context).colorScheme;
+    final textTheme = Theme.of(context).textTheme;
+    final l10n = AppLocalizations.of(context)!;
+
+    return Column(
+      crossAxisAlignment: CrossAxisAlignment.start,
+      children: [
+        Text(
+          l10n.stats_insights,
+          style: textTheme.headlineSmall?.copyWith(fontWeight: FontWeight.bold),
+        ),
+        SizedBox(height: spacing.sectionGap),
+        // Spending Prediction Banner
+        Consumer(
+          builder: (context, ref, child) {
+            final predictionAsync = ref.watch(predictedSpendingProvider);
+            return predictionAsync.when(
+              data: (predicted) {
+                if (predicted <= 0) return const SizedBox.shrink();
+                return Container(
+                  padding: EdgeInsets.all(spacing.cardInner),
+                  decoration: BoxDecoration(
+                    gradient: LinearGradient(
+                      colors: [
+                        color.primaryContainer,
+                        color.secondaryContainer,
+                      ],
+                    ),
+                    borderRadius: BorderRadius.circular(spacing.radiusMedium),
+                  ),
+                  child: Row(
+                    children: [
+                      Icon(
+                        LucideIcons.trendingUp,
+                        color: color.primary,
+                        size: 32,
+                      ),
+                      SizedBox(width: spacing.elementGap),
+                      Expanded(
+                        child: Column(
+                          crossAxisAlignment: CrossAxisAlignment.start,
+                          children: [
+                            Text(
+                              l10n.stats_nextMonthForecast,
+                              style: textTheme.labelLarge?.copyWith(
+                                color: color.onPrimaryContainer,
+                              ),
+                            ),
+                            SizedBox(height: spacing.elementGap),
+                            CurrencyText(
+                              amount: predicted,
+                              style: textTheme.headlineSmall?.copyWith(
+                                fontWeight: FontWeight.bold,
+                              ),
+                              compact: false,
+                              fixedLength: 0,
+                            ),
+                          ],
+                        ),
+                      ),
+                    ],
+                  ),
+                );
+              },
+              loading: () => const SizedBox.shrink(),
+              error: (_, __) => const InlineError(),
+            );
+          },
+        ),
+
+        // Insights Nudges
+        if (data.categoryData.isNotEmpty) ...[
+          Row(
+            children: [
+              Expanded(
+                child: _buildInsightCard(
+                  l10n.stats_topSpending,
+                  (data.categoryData.entries.toList()
+                        ..sort((a, b) => b.value.compareTo(a.value)))
+                      .first
+                      .key,
+                  LucideIcons.trendingUp,
+                  color,
+                  textTheme,
+                  spacing,
+                ),
+              ),
+              SizedBox(width: spacing.elementGap),
+              Expanded(
+                child: _buildInsightCard(
+                  l10n.statistics_dailyAverage,
+                  formatCurrency(
+                    data.avgDailySpend,
+                    code: BaseCurrency.code,
+                    decimals: 0,
+                  ),
+                  LucideIcons.calendar,
+                  color,
+                  textTheme,
+                  spacing,
+                ),
+              ),
+            ],
+          ),
+          SizedBox(height: spacing.elementGap),
+        ],
+
+        // Category Trends (Top 5)
+        Consumer(
+          builder: (context, ref, child) {
+            final categoryTrendsAsync = ref.watch(categoryTrendsProvider);
+            return categoryTrendsAsync.when(
+              data: (trends) {
+                if (trends.isEmpty) return const SizedBox.shrink();
+                final sortedTrends = trends.values.toList()
+                  ..sort((a, b) => b.thisMonth.compareTo(a.thisMonth));
+
+                return Card(
+                  elevation: 0,
+                  margin: const EdgeInsets.only(),
+                  color: color.surfaceContainerLow,
+                  shape: RoundedRectangleBorder(
+                    borderRadius: BorderRadius.circular(spacing.radiusMedium),
+                    side: BorderSide(
+                      color: color.outlineVariant.withValues(alpha: 0.5),
+                    ),
+                  ),
+                  child: Padding(
+                    padding: EdgeInsets.all(spacing.cardInner),
+                    child: Column(
+                      crossAxisAlignment: CrossAxisAlignment.start,
+                      children: [
+                        Text(
+                          l10n.stats_categoryTrends,
+                          style: textTheme.titleLarge?.copyWith(
+                            fontWeight: FontWeight.bold,
+                          ),
+                        ),
+                        const SizedBox(height: 20),
+                        ...sortedTrends.take(5).map(
+                              (trend) => Padding(
+                                padding: const EdgeInsets.only(bottom: 16),
+                                child: Column(
+                                  crossAxisAlignment: CrossAxisAlignment.start,
+                                  children: [
+                                    Row(
+                                      mainAxisAlignment:
+                                          MainAxisAlignment.spaceBetween,
+                                      children: [
+                                        Text(
+                                          trend.categoryName,
+                                          style: textTheme.bodyLarge?.copyWith(
+                                            fontWeight: FontWeight.w600,
+                                          ),
+                                        ),
+                                        Row(
+                                          children: [
+                                            CurrencyText(
+                                              amount: trend.thisMonth,
+                                              style: textTheme.titleSmall
+                                                  ?.copyWith(
+                                                fontWeight: FontWeight.bold,
+                                              ),
+                                            ),
+                                            if (trend.changePercent != 0) ...[
+                                              SizedBox(
+                                                width: spacing.elementGap,
+                                              ),
+                                              Container(
+                                                padding:
+                                                    const EdgeInsets.symmetric(
+                                                  horizontal: 8,
+                                                  vertical: 4,
+                                                ),
+                                                decoration: BoxDecoration(
+                                                  color:
+                                                      (trend.changePercent > 0
+                                                              ? color.error
+                                                              : color.primary)
+                                                          .withValues(
+                                                    alpha: 0.1,
+                                                  ),
+                                                  borderRadius:
+                                                      BorderRadius.circular(
+                                                    spacing.radiusMedium,
+                                                  ),
+                                                ),
+                                                child: Row(
+                                                  children: [
+                                                    Icon(
+                                                      trend.changePercent > 0
+                                                          ? LucideIcons.arrowUp
+                                                          : Icons
+                                                              .arrow_downward,
+                                                      size: 12,
+                                                      color:
+                                                          trend.changePercent >
+                                                                  0
+                                                              ? color.error
+                                                              : color.primary,
+                                                    ),
+                                                    const SizedBox(width: 2),
+                                                    Text(
+                                                      '${trend.changePercent.abs().toStringAsFixed(0)}%',
+                                                      style: textTheme.bodySmall
+                                                          ?.copyWith(
+                                                        color:
+                                                            trend.changePercent >
+                                                                    0
+                                                                ? color.error
+                                                                : color.primary,
+                                                        fontWeight:
+                                                            FontWeight.w600,
+                                                      ),
+                                                    ),
+                                                  ],
+                                                ),
+                                              ),
+                                            ],
+                                          ],
+                                        ),
+                                      ],
+                                    ),
+                                    SizedBox(height: spacing.elementGap),
+                                    _AnimatedMetricBar(
+                                      progress: (trend.thisMonth /
+                                              sortedTrends.first.thisMonth)
+                                          .clamp(0.0, 1.0),
+                                      barColor: color.primary,
+                                      bgColor: color.surfaceContainerHighest,
+                                    ),
+                                  ],
+                                ),
+                              ),
+                            ),
+                      ],
+                    ),
+                  ),
+                );
+              },
+              loading: () => const SizedBox.shrink(),
+              error: (_, __) => const InlineError(),
+            );
+          },
+        ),
+      ],
+    );
+  }
+
+  Widget _buildInsightCard(
+    String label,
+    String value,
+    IconData icon,
+    ColorScheme color,
+    TextTheme textTheme,
+    AppSpacing spacing,
+  ) {
+    return Card(
+      elevation: 0,
+      margin: EdgeInsets.only(top: spacing.elementGap),
+      color: color.surfaceContainerLow,
+      shape: RoundedRectangleBorder(
+        borderRadius: BorderRadius.circular(spacing.radiusMedium),
+        side: BorderSide(color: color.outlineVariant.withValues(alpha: 0.5)),
+      ),
+      child: Padding(
+        padding: EdgeInsets.all(spacing.cardInner),
+        child: Column(
+          crossAxisAlignment: CrossAxisAlignment.start,
+          children: [
+            Icon(icon, color: color.primary, size: 24),
+            SizedBox(height: spacing.elementGap),
+            Text(
+              label,
+              style: textTheme.bodySmall?.copyWith(
+                color: color.onSurfaceVariant,
+              ),
+            ),
+            SizedBox(height: spacing.elementGap),
+            Text(
+              value,
+              style: textTheme.titleMedium?.copyWith(
+                fontWeight: FontWeight.bold,
+              ),
+              maxLines: 1,
+              overflow: TextOverflow.ellipsis,
+            ),
+          ],
+        ),
+      ),
+    );
+  }
+}
+
+class _AnimatedMetricBar extends StatefulWidget {
+  final double progress;
+  final Color barColor;
+  final Color bgColor;
+
+  const _AnimatedMetricBar({
+    required this.progress,
+    required this.barColor,
+    required this.bgColor,
+  });
+
+  @override
+  State<_AnimatedMetricBar> createState() => _AnimatedMetricBarState();
+}
+
+class _AnimatedMetricBarState extends State<_AnimatedMetricBar>
+    with SingleTickerProviderStateMixin {
+  late AnimationController _ctrl;
+  late Animation<double> _anim;
+  bool _started = false;
+
+  @override
+  void initState() {
+    super.initState();
+    _ctrl = AnimationController(
+      vsync: this,
+      duration: const Duration(milliseconds: 1200),
+    );
+    _anim = CurvedAnimation(parent: _ctrl, curve: Curves.easeOutCubic);
+  }
+
+  @override
+  void dispose() {
+    _ctrl.dispose();
+    super.dispose();
+  }
+
+  @override
+  Widget build(BuildContext context) {
+    return VisibilityDetector(
+      key: ValueKey('metric_${widget.progress}_${widget.barColor.toARGB32()}'),
+      onVisibilityChanged: (info) {
+        if (info.visibleFraction > 0.3 && !_started) {
+          _started = true;
+          _ctrl.forward();
+        }
+      },
+      child: AnimatedBuilder(
+        animation: _anim,
+        builder: (_, __) => LinearProgressIndicator(
+          semanticsLabel: 'Progress',
+          value: _anim.value * widget.progress,
+          minHeight: 8,
+          backgroundColor: widget.bgColor,
+          valueColor: AlwaysStoppedAnimation(widget.barColor),
+        ),
+      ),
+    );
+  }
+}
