+++ lib/features/statistics/presentation/widgets/widgets.dart
@@ -1,3 +1,8 @@
 // Statistics Feature Widgets
 export 'category_pie_chart.dart';
+export 'financial_behavior_section.dart';
 export 'period_selector.dart';
+export 'spending_tags_section.dart';
+export 'statistics_chart_section.dart';
+export 'statistics_insights_section.dart';
+export 'statistics_metrics_section.dart';
